Sign-off + fixes (#1777)
* feat(backoffice-v2): added a mock version of the associated companies block

* Pre-requsite cells for associated companies (#1763)

* feat(backoffice-v2): created a re-usable dialog component

* refactor(backoffice-v2): created dumb cells for dialog and call to action, approve actions refactor

* fix(backoffice-v2): added missing size and variant to approve buttons

* feat(backoffice-v2): added initiate kyb button

* feat(backoffice-v2): added badge for associated companies block

* refactor(backoffice-v2): updated dialog styling

* feat(backoffice-v2): added a warning for when no document schemas were found

* feat(backoffice-v2): added warning for directors schemas

* fix(backoffice-v2): fixed additional properties condition and typo

* refactor(backoffice-v2): made refactors to allow creating document blocks with different data

* feat(backoffice-v2): now using next events as the condition for displaying the initiate kyb button

* refactor(*): sign-off + fixes
