+++ apps/backoffice-v2/src/pages/Entity/hooks/useTasks/utils/get-documents-schemas/get-documents-schemas.ts
@@ -1,0 +1,47 @@
+import {
+  getDocumentsByCountry,
+  getDocumentSchemaByCountry,
+  TAvailableDocuments,
+  TDocument,
+} from '@ballerine/common';
+import { TWorkflowById } from '@/domains/workflows/fetchers';
+
+export const getDocumentsSchemas = (
+  issuerCountryCode: Parameters<typeof getDocumentSchemaByCountry>[0],
+  workflow: TWorkflowById,
+) => {
+  if (!issuerCountryCode) return;
+
+  const documentSchemaByCountry = getDocumentSchemaByCountry(
+    issuerCountryCode,
+    workflow.workflowDefinition?.documentsSchema as TDocument[],
+  )
+    .concat(getDocumentsByCountry(issuerCountryCode))
+    .reduce((unique: TDocument[], item: TDocument) => {
+      const isDuplicate = unique.some(u => u.type === item.type && u.category === item.category);
+      if (!isDuplicate) {
+        unique.push(item);
+      }
+      return unique;
+    }, [] as TDocument[])
+    .filter((documentSchema: TDocument) => {
+      if (!workflow?.workflowDefinition?.config?.availableDocuments) return true;
+
+      const isIncludes = !!workflow?.workflowDefinition?.config?.availableDocuments.find(
+        (availableDocument: TAvailableDocuments[number]) =>
+          availableDocument.type === documentSchema.type &&
+          availableDocument.category === documentSchema.category,
+      );
+
+      return isIncludes;
+    });
+
+  if (!Array.isArray(documentSchemaByCountry) || !documentSchemaByCountry.length) {
+    console.warn(
+      `No document schema found for issuer country code of "${issuerCountryCode}" and documents schema of\n`,
+      workflow.workflowDefinition?.documentsSchema,
+    );
+  }
+
+  return documentSchemaByCountry;
+};
